+++ backend/src/main/kotlin/com/lute/db/repositories/StatusRepository.kt
@@ -10,11 +10,11 @@ import org.jetbrains.exposed.sql.transactions.transaction
 class StatusRepository {
   fun findAll(): List<Status> = transaction { StatusesTable.selectAll().map { it.toStatus() } }
 
-  fun findById(id: Int): Status? = transaction {
+  fun findById(id: Long): Status? = transaction {
     StatusesTable.selectAll().where { StatusesTable.StID eq id }.singleOrNull()?.toStatus()
   }
 
-  fun save(status: Status): Int = transaction {
+  fun save(status: Status): Long = transaction {
     StatusesTable.insert {
           it[StText] = status.text
           it[StAbbreviation] = status.abbreviation
